refac: move file reading to main
