+++ src/lib.rs
@@ -5,33 +5,6 @@ mod compiler;
 mod errors;
 mod parser;
 
-use anyhow::Result;
-use std::collections::HashSet;
-use std::path::PathBuf;
-
-pub fn get_file_paths(
-    root: PathBuf,
-    target_exts: &HashSet<&'static str>,
-) -> Result<Vec<PathBuf>> {
-    let mut result = Vec::<PathBuf>::new();
-    if root.is_dir() {
-        for entry in root.read_dir()?.flatten() {
-            let path = entry.path();
-            if path.is_dir() {
-                result.append(&mut get_file_paths(path, target_exts)?);
-            } else if let Some(ext) = path.extension() {
-                match ext.to_str() {
-                    Some(e) if target_exts.contains(e) => {
-                        result.push(path);
-                    }
-                    _ => (),
-                }
-            }
-        }
-    }
-    Ok(result)
-}
-
-pub fn python_to_mermaid(path: &'static str) -> Result<()> {
+pub fn python_to_mermaid(path: &'static str) -> Result<(), Box<dyn std::error::Error>> {
     todo!()
 }
